# docs/openapi-migrated-modules.yaml
openapi: 3.0.3
info:
  title: Shequyiyuan Migrated Modules API
  version: 1.1.0
  description: |
    覆盖当前已迁移模块（认证、健康检查、字典）的请求参数、响应结构与标准错误码。
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/configs:
    get: {tags: [config], summary: 配置列表, security: [{ bearerAuth: [] }], responses: {'200': {description: 查询成功}}}
    post: {tags: [config], summary: 新增配置, security: [{ bearerAuth: [] }], responses: {'200': {description: 新增成功}}}
  /api/v1/configs/{id}:
    get: {tags: [config], summary: 配置详情, security: [{ bearerAuth: [] }], responses: {'200': {description: 查询成功}}}
    put: {tags: [config], summary: 编辑配置, security: [{ bearerAuth: [] }], responses: {'200': {description: 更新成功}}}
  /api/v1/configs/{id}/status:
    patch: {tags: [config], summary: 配置状态变更, security: [{ bearerAuth: [] }], responses: {'200': {description: 状态变更成功}}}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...